refactoring and adding message ack
