+++ go.mod
@@ -8,5 +8,4 @@ require (
 	github.com/deckarep/golang-set/v2 v2.6.0 // indirect
 	github.com/go-jose/go-jose/v3 v3.0.3 // indirect
 	github.com/go-stack/stack v1.8.1 // indirect
-	github.com/valyala/fastjson v1.6.4 // indirect
 )
